user authorized.

+++ model/index.js
@@ -1,28 +1,29 @@
 const Contact = require("./schemas/contact");
 
-const listContacts = async () => {
+const listContacts = async (userId) => {
   try {
-    const results = await Contact.find({});
+    const results = await Contact.find({ owner: userId });
     return results;
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const getContactById = async (contactId) => {
+const getContactById = async (userId, contactId) => {
   try {
     const result = await Contact.findOne({
       _id: contactId,
-    });
+      owner: userId,
+    }).populate({ path: "owner" });
     return result;
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const removeContact = async (contactId) => {
+const removeContact = async (userId, contactId) => {
   try {
-    const result = Contact.findByIdAndRemove({ _id: contactId });
+    const result = Contact.findByIdAndRemove({ _id: contactId, owner: userId });
     return result;
   } catch (error) {
     console.log(error.message);
@@ -38,11 +39,12 @@ const addContact = async (body) => {
   }
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   try {
     const result = await Contact.findOneAndUpdate(
       {
         _id: contactId,
+        owner: userId,
       },
       { ...body },
       { new: true }
@@ -53,11 +55,12 @@ const updateContact = async (contactId, body) => {
   }
 };
 
-const updateStatusContact = async (contactId, body) => {
+const updateStatusContact = async (userId, contactId, body) => {
   try {
     const result = await Contact.findOneAndUpdate(
       {
         _id: contactId,
+        owner: userId,
       },
       { ...body },
       { new: true }
